Changes to address comments on PR #1554 from Feb 4th
