+++ package.json
@@ -46,7 +46,7 @@
 		"testData": "node -e 'require(\"./src/server/data/automatedTestingData.js\").insertSpecialUnitsConversionsMetersGroups()'",
 		"webData": "node -e 'require(\"./src/server/data/websiteData.js\").insertWebsiteData()'",
 		"addLogMsg": "node -e 'require(\"./src/server/services/addLogMsg.js\").addLogMsgToDB()'",
-		"changePostgresPassword": "node ./src/scripts/changePostgresPass.js"
+		"changePostgresPassword": "node -e 'require(\"./src/server/util/changePostgresPass\").changePasswords()'"
 	},
 	"nodemonConfig": {
 		"watch": [
